clear some errors and bugs

+++ frontend/src/components/Shop/AllEvents.jsx
@@ -1,6 +1,5 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { deleteProduct } from "../../redux/actions/product";
 import { AiOutlineDelete, AiOutlineEye } from "react-icons/ai";
 import { Link } from "react-router-dom";
 import Loader from "../Layout/Loader";
@@ -9,34 +8,28 @@ import { DataGrid } from "@mui/x-data-grid";
 import { deleteEvent, getAllEventsShop } from "../../redux/actions/event";
 
 const AllEvents = () => {
+  const dispatch = useDispatch();
+
   const { events, isLoading } = useSelector((state) => state.events);
   const { seller } = useSelector((state) => state.seller);
 
-  const dispatch = useDispatch();
-
   useEffect(() => {
-    dispatch(getAllEventsShop(seller?._id));
-  }, [dispatch]);
+    if (seller?._id) {
+      dispatch(getAllEventsShop(seller._id));
+    }
+  }, [dispatch, seller]);
 
   const handleDelete = (id) => {
-    dispatch(deleteEvent(id));
-    window.location.reload();
+    if (window.confirm("Are you sure you want to delete this event?")) {
+      dispatch(deleteEvent(id));
+      dispatch(getAllEventsShop(seller._id));
+    }
   };
 
   const columns = [
-    { field: "id", headerName: "Product Id", minWidth: 150, flex: 0.7 },
-    {
-      field: "name",
-      headerName: "Name",
-      minWidth: 180,
-      flex: 1.4,
-    },
-    {
-      field: "price",
-      headerName: "Price",
-      minWidth: 100,
-      flex: 0.6,
-    },
+    { field: "id", headerName: "Event Id", minWidth: 150, flex: 0.7 },
+    { field: "name", headerName: "Name", minWidth: 180, flex: 1.4 },
+    { field: "price", headerName: "Price", minWidth: 100, flex: 0.6 },
     {
       field: "Stock",
       headerName: "Stock",
@@ -44,64 +37,49 @@ const AllEvents = () => {
       minWidth: 80,
       flex: 0.5,
     },
-
     {
       field: "sold",
-      headerName: "Sold out",
+      headerName: "Sold Out",
       type: "number",
       minWidth: 130,
       flex: 0.6,
     },
     {
       field: "Preview",
-      flex: 0.8,
+      headerName: "Preview",
       minWidth: 100,
-      headerName: "",
-      type: "number",
+      flex: 0.8,
       sortable: false,
-      renderCell: (params) => {
-        return (
-          <>
-            <Link to={`/product/${params.id}`}>
-              <Button onClick={() => handleDelete(params.id)}>
-                <AiOutlineEye size={20} />
-              </Button>
-            </Link>
-          </>
-        );
-      },
+      renderCell: (params) => (
+        <Link to={`/product/${params.row.id}`}>
+          <Button>
+            <AiOutlineEye size={20} />
+          </Button>
+        </Link>
+      ),
     },
     {
       field: "Delete",
-      flex: 0.8,
+      headerName: "Delete",
       minWidth: 120,
-      headerName: "",
-      type: "number",
+      flex: 0.8,
       sortable: false,
-      renderCell: (params) => {
-        return (
-          <>
-            <Button>
-              <AiOutlineDelete size={20} />
-            </Button>
-          </>
-        );
-      },
+      renderCell: (params) => (
+        <Button onClick={() => handleDelete(params.row.id)}>
+          <AiOutlineDelete size={20} />
+        </Button>
+      ),
     },
   ];
 
-  const row = [];
-
-  events &&
-    events.forEach((item) => {
-      row.push({
-        id: item._id,
-        name: item.name,
-        price: "US$ " + item.discountPrice,
-        Stock: item.stock,
-        sold: 10,
-      });
-    });
+  const rows =
+    events?.map((item) => ({
+      id: item._id,
+      name: item.name,
+      price: "US$ " + item.discountPrice,
+      Stock: item.stock,
+      sold: item.sold || 0,
+    })) || [];
 
   return (
     <>
@@ -110,7 +88,7 @@ const AllEvents = () => {
       ) : (
         <div className="w-full mx-8 pt-1 mt-10 bg-white">
           <DataGrid
-            rows={row}
+            rows={rows}
             columns={columns}
             pageSize={10}
             disableSelectionOnClick
